Reorganize key cracker into own class
