+++ lib/message_shift.rb
@@ -68,16 +68,4 @@ class MessageShift
     find_char(new_value)
   end
 
-  def find_crack_key(date)
-    reversed_message = @message.chars.reverse
-    message_length = @message.length % 4
-
-    offset = ShiftCreator.square_number_last_four(date)
-    key_hash = {}
-    key_hash["A"] = (find_index(reversed_message[0]) - find_index("d")) #- #offset
-    key_hash["B"] = (find_index(reversed_message[1]) - find_index("n")) #- #offset
-    key_hash["C"] = (find_index(reversed_message[2]) - find_index("e")) #- #offset
-    key_hash["D"] = (find_index(reversed_message[3]) - find_index(" ")) #- #offset
-    key_hash
-  end
 end
